Logging improvements. Closes #9
With the customized logging output for parallel runs, it is now possible to export all logs to disk. It will log to a timestamped directory for each parallel run.

The logging format for disk is:

    [IS0-8601 DATE] LEVEL   message

While the logging format for stdout is:

    [HOST] [IS0-8601 DATE] LEVEL   message

The reason they are different is because the files themselves contain the hostname, so there is no need to prepend it. Stdout/Stderr will also use colors in output, whereas this is omitted in the logfiles for clarity.

Note that chef_solo_cup will attempt to create all missing directories.